Refactor into a presentation node root
