+++ src/app/progress/Progress.tsx
@@ -28,7 +28,7 @@ import { characterComponentSortSelector } from '../settings/character-sort';
 import { Subscriptions } from '../rx-utils';
 import { refresh$ } from '../shell/refresh';
 import CollapsibleTitle from '../dim-ui/CollapsibleTitle';
-import PresentationNode, { countCollectibles } from '../collections/PresentationNode';
+import PresentationNodeRoot from '../collections/PresentationNodeRoot';
 
 const factionOrder = [
   611314723, // Vanguard,
@@ -144,11 +144,6 @@ class Progress extends React.Component<Props, State> {
       firstCharacterProgression[2772425241]
     ];
 
-    // TODO: how to search and find an item within all nodes?
-    const fullNodePath: number[] = [];
-    fullNodePath.unshift(1024788583);
-    const collectionCounts = countCollectibles(defs, 1024788583, profileInfo);
-
     const profileMilestonesContent = (profileMilestones.length > 0 || profileQuests.length > 0) && (
       <>
         <div className="profile-content">
@@ -219,17 +214,10 @@ class Progress extends React.Component<Props, State> {
           </div>
         </div>
         <div className="section">
-          <h3 className="category-title">{t('Vendors.Collections')}</h3>
-          <PresentationNode
-            collectionCounts={collectionCounts}
+          <PresentationNodeRoot
             presentationNodeHash={1024788583}
             defs={defs}
             profileResponse={profileInfo}
-            buckets={this.state.progress.buckets}
-            ownedItemHashes={new Set()}
-            path={fullNodePath}
-            onNodePathSelected={(nodePath: number[]) => {}}
-            parents={[]}
           />
         </div>
         <hr />
